Moved Kernel class to its own file
