+++ archey/kernel.py
@@ -1,0 +1,9 @@
+from subprocess import check_output
+
+
+class Kernel:
+    def __init__(self):
+        self.value = check_output(
+            ['uname', '-r'],
+            universal_newlines=True
+        ).rstrip()
